requirements.txt
pandas==2.2.2
requests==2.32.3
numpy==2.0.2
fastapi[standard]==0.115.6
pydantic==2.10.5
uvicorn==0.34.0
python-dotenv==1.0.1
pydantic-settings==2.10.1
pymongo==4.6.1
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
bson==0.5.10
oracledb==2.1.2
openpyxl==3.1.5
boto3==1.40.34